+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Yorc Changelog
 
+## UNRELEASED
+
 ## 4.0.0-rc.1 (March 30, 2020)
 
 ### FEATURES