+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### DEPENDENCIES
 
-* The orchestrator requires now at least Terraform google plugin with version constraint `~ 2.0` ([GH-628](https://github.com/ystia/yorc/issues/628))
+* The orchestrator requires now at least Terraform google plugin with version constraint `~ 2.20` ([GH-628](https://github.com/ystia/yorc/issues/628))
 
 ### BUG FIXES